feat: add expr interval support
